% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_conversion.R
\name{get_conversion_factors}
\alias{get_conversion_factors}
\title{Get Internal Conversion Factors Database}
\usage{
get_conversion_factors()
}
\value{
Data frame with conversion factors
}
\description{
Returns the internal database of biomass conversion factors for benthic taxa.
}
\details{
This is an internal function that stores the conversion factors database.
Users typically don't need to call this directly - it's used automatically
by ConvertBiomass() when factors = NULL.
}
\keyword{internal}
